fix: correct syntax for if statement in script
The previous version of the script used an incorrect syntax for the if statement, which caused the script to fail when executed. This commit fixes the issue by using double brackets [[ and ]] to properly compare the value of the $updates variable, ensuring that the script runs as intended. Icon added.