chore(add): alacritty
I have decided to make Alacritty my primary terminal once again. The main reason for this switch is the smoother handling of scrolling and marking longer text simultaneously. Additionally, Alacritty now utilizes TOML instead of YAML for its configuration, a change that I find enjoyable to explore and experiment with.